+++ src/test/java/com/bulain/mybatis/service/UserServiceImplTest.java
@@ -21,7 +21,7 @@ public class UserServiceImplTest extends ServiceTestCase {
 	
 	@Test
 	public void testGet() {
-		User user = userService.get(Integer.valueOf(102));
+		User user = userService.get(Long.valueOf(102));
 		assertNotNull(user);
 		
 		assertEquals("first_name_102", user.getFirstName());
@@ -39,7 +39,7 @@ public class UserServiceImplTest extends ServiceTestCase {
 	@Test
 	public void testUpdate() {
 		User user = new User();
-		user.setId(Integer.valueOf(104));
+		user.setId(Long.valueOf(104));
 		user.setFirstName("firstName-updated");
 		user.setLastName("lastName-updated");
 		userService.update(user, true);
@@ -47,7 +47,7 @@ public class UserServiceImplTest extends ServiceTestCase {
 	
 	@Test
 	public void testDelete() {
-		userService.delete(Integer.valueOf(101));
+		userService.delete(Long.valueOf(101));
 	}
 
 	@Test
